refactor(emitter): split portability symbol helpers
